Move Entity and EntityBuilder to be pure interfaces
