+++ sdk/common/src/main/java/io/opentelemetry/sdk/resources/internal/Entity.java
@@ -5,7 +5,6 @@
 
 package io.opentelemetry.sdk.resources.internal;
 
-import com.google.auto.value.AutoValue;
 import io.opentelemetry.api.common.Attributes;
 import javax.annotation.Nullable;
 import javax.annotation.concurrent.Immutable;
@@ -27,8 +26,7 @@ import javax.annotation.concurrent.Immutable;
  * at any time.
  */
 @Immutable
-@AutoValue
-public abstract class Entity {
+public interface Entity {
   /**
    * Returns the entity type string of this entity. Must not be null.
    *
@@ -61,37 +59,16 @@ public abstract class Entity {
   public abstract String getSchemaUrl();
 
   /**
-   * Returns a {@link Entity}.
-   *
-   * @param entityType the entity type string of this entity.
-   * @param id a map of attributes that identify the entity.
-   * @param description a map of attributes that describe the entity.
-   * @return a {@code Entity}.
-   * @throws NullPointerException if {@code id} or {@code description} is null.
-   * @throws IllegalArgumentException if entityType string, attribute key or attribute value is not
-   *     a valid printable ASCII string or exceed {@link AttributeCheckUtil#MAX_LENGTH} characters.
-   */
-  static final Entity create(
-      String entityType, Attributes id, Attributes description, @Nullable String schemaUrl) {
-    AttributeCheckUtil.isValid(entityType);
-    AttributeCheckUtil.checkAttributes(id);
-    AttributeCheckUtil.checkAttributes(description);
-    return new AutoValue_Entity(entityType, id, description, schemaUrl);
-  }
-
-  /**
    * Returns a new {@link EntityBuilder} instance populated with the data of this {@link Entity}.
    */
-  public final EntityBuilder toBuilder() {
-    return new EntityBuilder(this);
-  }
+  EntityBuilder toBuilder();
 
   /**
    * Returns a new {@link EntityBuilder} instance for creating arbitrary {@link Entity}.
    *
    * @param entityType the entity type string of this entity.
    */
-  public static final EntityBuilder builder(String entityType) {
-    return new EntityBuilder(entityType);
+  public static EntityBuilder builder(String entityType) {
+    return SdkEntity.builder(entityType);
   }
 }
